Domain Model finished
Domain Model finished and tested.
Added UserAccount support, GrantedAuthorities and UserDetails for Login functionality.

Also, LoginService for getting the LoogedIn user and test the different functionalities.

Admin Service and Repo added. Take a look for the findByPrincipal method. It will be used for asserting that the different users(Roles) dont access functions not of them